[SDK-258] Added Full Name and AgeVerified attributes
