+++ src/Example/Controllers/HomeController.cs
@@ -8,19 +8,8 @@ namespace Example.Controllers
     [Authorize]
     public class HomeController : Controller
     {
-        public static byte[] PhotoBytes { get; set; }
-
         public ActionResult Index()
         {
-            var user = GetUser();
-
-            if (user.Base64Photo == null)
-                ViewBag.Message = "No photo provided, change the application settings to request a photo from the user for this demo";
-            else
-            {
-                ViewBag.Photo = user.Base64Photo;
-                PhotoBytes = user.Photo;
-            }
             return View();
         }
 
@@ -30,14 +19,6 @@ namespace Example.Controllers
             int userId = int.Parse(claimsIdentity.FindFirst(ClaimTypes.Name).Value);
 
             return UserManager.GetUserById(userId);
-        }
-
-        public FileContentResult DownloadImageFile()
-        {
-            if (PhotoBytes == null)
-                throw new InvalidOperationException("The 'PhotoBytes' variable has not been set");
-
-            return File(PhotoBytes, System.Net.Mime.MediaTypeNames.Application.Octet, "YotiSelfie.jpg");
         }
     }
 }
