+++ src/Example/Views/Account/Connect.cshtml
@@ -1,6 +1,118 @@
 ﻿@{
     Layout = "~/Views/Shared/_LayoutSimple.cshtml";
 }
+@model Example.Models.User
+
 <div class="jumbotron">
     @ViewBag.Error
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            <div class="thumbnail">
+                @if (Model != null)
+                {
+                    if (Model.Base64Photo == null)
+                    {
+                        <p>@ViewBag.Message</p>
+                    }
+                    else
+                    {
+                        <img src="@Model.Base64Photo" />
+                        @Html.ActionLink(
+                             linkText: "Download Image",
+                             actionName: "DownloadImageFile",
+                             controllerName: "Account")
+                    }
+                }
+            </div>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            ID:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.Id)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Full Name:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.FullName)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Given Names:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.GivenNames)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Family Name:
+        </div>
+
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.FamilyName)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Mobile Number:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.MobileNumber)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Email Address:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.EmailAddress)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Date of Birth:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.DateOfBirth)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Age Verified:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.IsAgeVerified)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Address:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.Address)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Gender:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.Gender)
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            Nationality:
+        </div>
+        <div class="col-md-4 col-sm-6 col-xs-12">
+            @Html.DisplayFor(m => m.Nationality)
+        </div>
+    </div>
 </div>
