+++ src/Example/Views/Home/Index.cshtml
@@ -1,30 +1,1 @@
-﻿@{
-    ViewBag.Title = "Home Page";
-}
-
-<div class="jumbotron">
-    <div class="row">
-        <div class="col-md-4 col-sm-6 col-xs-12">
-            <div class="thumbnail">
-                @if (ViewBag.Photo == null)
-                {
-                    <p>@ViewBag.Message</p>
-                }
-                else
-                {
-                    <img src="@Html.Raw(ViewBag.Photo)" alt="Responsive image">
-
-                    @Html.ActionLink(
-                        linkText: "Download Image",
-                        actionName: "DownloadImageFile",
-                        controllerName: "Home")
-                }
-            </div>
-        </div>
-        <div class="col-md-8 col-sm-6 col-xs-12">
-            <h1>Hello</h1>
-            <p class="lead">I was only kidding about the world's finest example project but since you are here you might as well check out Yoti's fabulous website</p>
-            <p><a href="http://yoti.com" class="btn btn-primary btn-lg">Learn more &raquo;</a></p>
-        </div>
-    </div>
-</div>
+﻿
